# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agenda+ API",
        "description": "API RESTful para sistema de gestão de consultas médicas Agenda+",
        "contact": {
            "name": "Suporte Agenda+",
            "email": "suporte@agendaplus.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor da API"
        }
    ],
    "paths": {
...
        "/doctor/schedules": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Listar horários da agenda",
                "description": "Lista todos os horários configurados pelo médico autenticado",
                "operationId": "a97af96e73490e863820b21ec6222ed5",
                "responses": {
                    "200": {
                        "description": "Lista de horários paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Criar horário na agenda",
                "description": "Adiciona um novo horário na agenda do médico. A agenda deve ter no mínimo 4h semanais disponíveis.",
                "operationId": "7c190cba42099b2fd5e644c018394335",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "day_of_week",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "day_of_week": {
                                        "description": "1=Segunda, 7=Domingo",
                                        "type": "integer",
                                        "maximum": 7,
                                        "minimum": 1,
                                        "example": 1
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "12:00"
                                    },
                                    "slot_duration_minutes": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "is_blocked": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Horário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Conflito de horário ou mínimo de 4h não atingido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/doctor/schedules/{id}": {
            "put": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Atualizar horário",
                "description": "Atualiza um horário existente. Não é possível atualizar horários que possuem consultas agendadas.",
                "operationId": "cbb7992417631c4c0d8a02639ff1c374",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "day_of_week": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "13:00"
                                    },
                                    "is_blocked": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Horário atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Horário possui consultas agendadas ou mínimo de 4h não atingido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Remover horário",
                "description": "Remove um horário da agenda. Não é possível remover horários que possuem consultas agendadas. A agenda deve manter no mínimo 4h semanais.",
                "operationId": "97fe045cc82e710c2b7f51d2be975271",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Horário removido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Horário possui consultas agendadas ou mínimo de 4h não seria atingido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação Sanctum",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}